+++ packages/web-component/src/lib/descope-wc/BaseDescopeWc.ts
@@ -19,6 +19,7 @@ import {
   DebugState,
   FlowState,
   FlowStateUpdateFn,
+  ILogger,
   SdkConfig,
   ThemeOptions,
 } from '../types';
@@ -29,6 +30,8 @@ declare const BUILD_VERSION: string;
 
 // this base class is responsible for WC initialization
 class BaseDescopeWc extends HTMLElement {
+  logger: ILogger = console;
+
   static get observedAttributes() {
     return [
       'project-id',
@@ -36,6 +39,7 @@ class BaseDescopeWc extends HTMLElement {
       'base-url',
       'tenant',
       'theme',
+      'locale',
       'debug',
       'telemetryKey',
       'redirect-url',
@@ -53,6 +57,19 @@ class BaseDescopeWc extends HTMLElement {
   };
 
   #init = false;
+
+  loggerWrapper = {
+    error: (message: string, description = '') => {
+      this.logger.error(message, description, new Error());
+      this.#updateDebuggerMessages(message, description);
+    },
+    warn: (message: string, description = '') => {
+      this.logger.warn(message, description);
+    },
+    info: (message: string, description = '', state: any = {}) => {
+      this.logger.info(message, description, state);
+    },
+  };
 
   #flowState = new State<FlowState>({ deferredRedirect: false } as FlowState);
 
@@ -129,6 +146,10 @@ class BaseDescopeWc extends HTMLElement {
     return this.getAttribute('debug') === 'true';
   }
 
+  get locale() {
+    return this.getAttribute('locale') || undefined;
+  }
+
   get theme(): ThemeOptions {
     const theme = this.getAttribute('theme') as ThemeOptions;
 
@@ -164,6 +185,7 @@ class BaseDescopeWc extends HTMLElement {
       'base-url',
       'tenant',
       'theme',
+      'locale',
       'debug',
       'telemetryKey',
       'redirect-url',
@@ -259,7 +281,7 @@ class BaseDescopeWc extends HTMLElement {
         executionContext: { geo: headers['x-geo'] },
       };
     } catch (e) {
-      this.logger.error(
+      this.loggerWrapper.error(
         'Cannot get config file',
         'make sure that your projectId & flowId are correct'
       );
@@ -287,7 +309,7 @@ class BaseDescopeWc extends HTMLElement {
       const { body } = await fetchContent(themeUrl, 'text');
       styleEle.innerText = body;
     } catch (e) {
-      this.logger.error(
+      this.loggerWrapper.error(
         'Cannot fetch theme file',
         'make sure that your projectId & flowId are correct'
       );
@@ -350,17 +372,12 @@ class BaseDescopeWc extends HTMLElement {
     return config;
   }
 
-  logger = {
-    error: (message: string, description = '') => {
-      // eslint-disable-next-line no-console
-      console.error(message, description, new Error());
-      this.#updateDebuggerMessages(message, description);
-    },
-    info: (message: string, description = '') => {
-      // eslint-disable-next-line no-console
-      console.log(message, description);
-    },
-  };
+  async getTargetLocales() {
+    const flowConfig = await this.getFlowConfig();
+    return (flowConfig?.targetLocales || []).map((locale: string) =>
+      locale.toLowerCase()
+    );
+  }
 
   #handleKeyPress() {
     // we want to simulate submit when the user presses Enter
@@ -434,6 +451,7 @@ class BaseDescopeWc extends HTMLElement {
         baseUrl: this.baseUrl,
         tenant: this.tenant,
         redirectUrl: this.redirectUrl,
+        locale: this.locale,
         stepId,
         executionId,
         token,