+++ packages/web-js-sdk/src/sdk/flow.ts
@@ -21,7 +21,8 @@ export default (coreSdk: CoreSdk) => ({
   start: async (...args: ReplaceParam<CoreSdkFlowStartArgs, '1', Options>) => {
     const webAuthnSupport = await isSupported();
     const decoratedOptions = {
-      redirectUrl: window.location.href,
+      redirectUrl: window.location.href, // for backward compatibility
+      location: window.location.href,
       ...args[1],
       deviceInfo: {
         webAuthnSupport,