+++ packages/web-js-sdk/test/sdk.test.ts
@@ -23,6 +23,7 @@ describe('sdk', () => {
     expect(JSON.parse(mockFetch.mock.calls[0][1].body)).toMatchObject({
       options: {
         redirectUrl: 'http://localhost/',
+        location: 'http://localhost/',
         deviceInfo: { webAuthnSupport: false },
       },
       flowId: 'id',